requirements.txt
PyHamcrest==2.0.4
requests==2.27.1
Appium-Python-Client==0.48
selenium==3.141.0
JPype1==0.7.5
paramiko==2.4.3
Pillow~=10.4.0
scikit-image~=0.24.0
opencv-python~=4.10.0.84
ujson==4.3.0
locust==1.6.0
mitmproxy==5.0.1
pycryptodome==3.9.8
python-binary-memcached==0.30.1
kazoo==2.8.0
websockets==9.1
Js2Py==0.71
# pytest
pytest==5.0.0
pytest-xdist==1.30.0
allure-pytest==2.7.0
pytest-rerunfailures==8.0
pytest-collect-formatter2==0.1.3
# database
cx_Oracle==6.3.1
PyMySQL==1.0.2
redis==3.5.3
sqlacodegen==3.0.0b2
SQLAlchemy==1.4.39
loguru~=0.7.2
PyYAML~=6.0.2